+++ src/UI/components/HomeCategoryNav.jsx
@@ -1,31 +1,22 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
 import {
   StHomeCategoryNav,
   StHomeCategoryNavButton,
 } from '../styles/HomeCategoryNav.styled';
-
-import { useDispatch } from 'react-redux';
-import { getStudyType } from 'src/redux/modules/studyHandler';
+import useHomeAddress from 'src/hooks/useHomeAddress';
 
 function HomeCategoryNav() {
-  const [address, setAddress] = useState('all');
-
-  const onClickOnlineHandler = () => setAddress('online');
-  const onClickOfflineHandler = () => setAddress('offline');
-
-  const dispatch = useDispatch(state => state);
-
-  useEffect(() => {
-    dispatch(getStudyType(address));
-  }, [address]);
+  const [allHandler, onlineHandler, offlineHandler] = useHomeAddress();
 
   return (
     <StHomeCategoryNav>
-      {/* 보더 2px 늘어나는 거 생각해서 수정해라 */}
-      <StHomeCategoryNavButton onClick={onClickOfflineHandler}>
+      <StHomeCategoryNavButton onClick={allHandler}>
+        전체
+      </StHomeCategoryNavButton>
+      <StHomeCategoryNavButton onClick={onlineHandler}>
         지역
       </StHomeCategoryNavButton>
-      <StHomeCategoryNavButton onClick={onClickOnlineHandler}>
+      <StHomeCategoryNavButton onClick={offlineHandler}>
         온라인
       </StHomeCategoryNavButton>
     </StHomeCategoryNav>
